resolved some of the comments on pr review
