+++ frontend/src/communications/userApi.js
@@ -6,7 +6,6 @@ export const getUsers = async () => {
     const response = await api.get('/users')
     return response.data.users
   } catch (error) {
-    console.log(error)
     throw new Error('Failed to get users.')
   }
 }
@@ -39,7 +38,6 @@ export const getAllProducts = async () => {
     const response = await api.get('/products')
     return response.data.products
   } catch (error) {
-    console.log(error)
     throw new Error('Failed to get products.')
   }
 }
@@ -50,7 +48,6 @@ export const getProductsByUserId = async (userId) => {
 
     return response.data.products
   } catch (error) {
-    console.log(error)
     throw new Error('Failed to get products.')
   }
 }
