+++ frontend/src/components/listOfProducts/ListHeader.js
@@ -12,7 +12,7 @@ const ListHeader = () => {
         <th scope="col">{t('product.city')}</th>
         <th scope="col">{t('product.category')}</th>
         <th scope="col">{t('product.owner')}</th>
-        <th>{t('')}</th>
+        <th></th>
       </tr>
     </thead>
   )
