+++ frontend/src/components/listofusers/ListHeader.js
@@ -31,7 +31,7 @@ const ListHeader = ({ onColumnClick, isSortAscending, sortKey }) => {
           {SortingSign('dateOfRegistration')}
         </th>
         <th> {t('userListHeader.actions')}</th>
-        <th>{t('')}</th>
+        <th></th>
       </tr>
     </thead>
   )
